+++ src/pages/home/music/Audiovisual.js
@@ -6,7 +6,7 @@ const Audiovisual = ({ id, openTab, ...props }) => {
     const { getPostsById } = useContextInfo();
 
     return (
-        <ExpandableSection {...props} id={id} title="Audiovisual" isFirstItem openTab={openTab}>
+        <ExpandableSection {...props} id={id} title="Audiovisual" openTab={openTab}>
             <div className="h-full w-full overflow-y-scroll">
                 <VideosList videos={getPostsById(id)} />
             </div>